// src/app/interfaces/news.interface.ts
// Generated by https://quicktype.io

export interface News {
  status: string;
  totalResults: number;
  articles: Article[];
}

export interface Article {
  source: Source;
  author: string;
  title: string;
  description: string;
  url: string;
  urlToImage: null | string;
  publishedAt: string;
  content: null | string;
}

export interface Source {
  id: null | string;
  name: string;
}

export interface ArticlesByCategoryAndPage {
  [key: string]: {
    page: number;
    articles: Article[];
  };
}